backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
arize-phoenix>=0.1.0